# docs/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PatchRequest": {
        "properties": {
          "coverage_half_extent": {
            "default": 4.5,
            "description": "Half side length of canonical axis-aligned square that must be fully covered before masking.",
            "exclusiveMinimum": 0.0,
            "title": "Coverage Half Extent",
            "type": "number"
          },
          "force_svg_large": {
            "default": false,
            "title": "Force Svg Large",
            "type": "boolean"
          },
          "formats": {
            "items": {
              "type": "string"
            },
            "title": "Formats",
            "type": "array"
          },
          "mask": {
            "additionalProperties": true,
            "title": "Mask",
            "type": "object"
          },
          "patch_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patch Version"
          },
          "png_height_px": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Png Height Px"
          },
          "png_width_px": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Png Width Px"
          },
          "retention": {
            "default": "centroid",
            "enum": [
              "centroid",
              "intersection",
              "clip"
            ],
            "title": "Retention",
            "type": "string"
          },
          "rotation_deg": {
            "default": 0.0,
            "title": "Rotation Deg",
            "type": "number"
          },
          "scale": {
            "default": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Scale",
            "type": "number"
          },
          "seed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "stl_extrusion_mm": {
            "default": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Stl Extrusion Mm",
            "type": "number"
          },
          "substitution_iterations": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Substitution Iterations"
          },
          "svg_deterministic_palette": {
            "default": false,
            "title": "Svg Deterministic Palette",
            "type": "boolean"
          },
          "svg_fill": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Svg Fill"
          },
          "svg_opacity": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Svg Opacity"
          },
          "svg_stroke": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Svg Stroke"
          },
          "svg_stroke_width": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Svg Stroke Width"
          },
          "tile_family": {
            "const": "spectre_tile_1_1",
            "default": "spectre_tile_1_1",
            "title": "Tile Family",
            "type": "string"
          },
          "tx": {
            "default": 0.0,
            "title": "Tx",
            "type": "number"
          },
          "ty": {
            "default": 0.0,
            "title": "Ty",
            "type": "number"
          }
        },
        "required": [
          "mask"
        ],
        "title": "PatchRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Deterministic Tier-1 API for Spectre / Tile(1,1) patches + masking + exporters. Attribution docs live in packaged `docs/ATTRIBUTION.md`.",
    "title": "Spectre Patch API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/capabilities": {
      "get": {
        "operationId": "caps_v1_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Caps"
      }
    },
    "/v1/downloads/{job_id}/{filename}": {
      "get": {
        "operationId": "signed_download_v1_downloads__job_id___filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "exp",
            "required": true,
            "schema": {
              "title": "Exp",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sig",
            "required": true,
            "schema": {
              "title": "Sig",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signed Download"
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/v1/jobs/{job_id}/urls": {
      "get": {
        "operationId": "signed_url_bundle_v1_jobs__job_id__urls_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ttl_seconds",
            "required": false,
            "schema": {
              "default": 900,
              "title": "Ttl Seconds",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Signed Url Bundle V1 Jobs  Job Id  Urls Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signed Url Bundle"
      }
    },
    "/v1/patch": {
      "post": {
        "operationId": "post_patch_v1_patch_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Patch V1 Patch Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Patch"
      }
    }
  },
  "tags": [
    {
      "name": "capabilities"
    },
    {
      "name": "jobs"
    }
  ]
}